idk why float works but int does not. quite funny actualy. but hey if it works dont touch it

+++ Silk.Net.OpenGLtest1/shaders/VertexShader.glsl
@@ -1,18 +1,18 @@
-#version 330 core
+#version 410 core
 layout (location = 0) in vec3 aPosition;
-layout (location = 1) in int blockID;
+layout (location = 1) in float blockID;
 
 uniform mat4 perspective;
 uniform mat4 projection;
 
 out DATA
 {
-	uint blockID;
+	int blockID;
 } data_out;
     
 void main() 
 {
     gl_Position = vec4(aPosition.xyz,1.0);
-    data_out.blockID = uint(blockID);
+    data_out.blockID = int(blockID);
 
 }
